+++ src/BasicMath.sol
@@ -2,22 +2,16 @@
 pragma solidity ^0.8.13;
 
 contract BasicMath {
-    function adder(
-        uint _a,
-        uint _b
-    ) external pure returns (uint sum, bool error) {
+    function adder(uint256 _a, uint256 _b) external pure returns (uint256 sum, bool error) {
         // Check for overflow
-        if (_a > type(uint).max - _b) {
+        if (_a > type(uint256).max - _b) {
             return (0, true);
         }
         sum = _a + _b;
         return (sum, false);
     }
 
-    function subtractor(
-        uint _a,
-        uint _b
-    ) external pure returns (uint difference, bool error) {
+    function subtractor(uint256 _a, uint256 _b) external pure returns (uint256 difference, bool error) {
         // Check for underflow
         if (_a < _b) {
             return (0, true);
